requirements/_base.txt
# dependencies used in all environments

# Non-Python dependencies
# sudo apt-get install python2.7, supervisor

                                   # LICENSE  Python3
                                   
bpython                            # MIT
django                             # MIT      y
gunicorn                           # MIT      y
djangorestframework                # BSD      y
